resume-ranker-service/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
PyPDF2==3.0.1
sentence-transformers>=2.5.0
scikit-learn==1.3.2
google-generativeai==0.3.2
numpy==1.24.3
huggingface-hub>=0.20.0
python-multipart==0.0.6

